+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/RequestError.kt
@@ -5,7 +5,7 @@ import com.soywiz.klock.DateTime
 sealed class RequestError
 
 data class RetryAfterError(
-    val seconds: Long,
+    val seconds: Seconds,
     val startCountingMillis: Long
 ) : RequestError() {
     val canContinue = (seconds * 1000L) + startCountingMillis