+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # TelegramBotAPI changelog
 
+## 0.30.3
+
+* `Common`:
+    * `Version`:
+        * `MicroUtils`: `0.3.0` -> `0.3.1`
+* `Core`:
+    * New type of requests exceptions `TooMuchRequestsException`. In fact it will be rare case when you will get this
+    exception
+    * `EmptyLimiter` has been renamed to `ExceptionsOnlyLimiter` and currently will stop requests after
+    `TooMuchRequestsException` happen until retry time is actual
+        * Now `ExceptionsOnlyLimiter` (previously `EmptyLimiter`) is a class
+    * `AbstractRequestCallFactory` currently will not look at the response and wait if it have `RetryAfter` error. New
+    behaviour aimed on delegating of this work to `RequestsLimiter`
+
 ## 0.30.2
 
 * `Common`: